requirements.txt
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
async-timeout==5.0.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.3
cachetools==6.2.1
celery==5.3.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.44.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
